stf-connect-class
